+++ src/main/java/com/motwg/shared/Task.java
@@ -3,6 +3,6 @@ package com.motwg.shared;
 import java.util.List;
 
 public interface Task {
-    public int partOne(List<String> input);
-    public int partTwo(List<String> input);
+    public long partOne(List<String> input);
+    public long partTwo(List<String> input);
 }
